v1.4.0.0 - All blocks without MN reward consensus
This version creates a fork at block 675000 that removes the coded enforcement of MN payments each block. Spork 8 is re-enabled to handle these and a warning is written to the log and block created without a MN payment if consensus can't be reached.

Version changed to 1.4.0.0
Protocol version incremented to 70006
Minimum peer version increased to 70005